Change "to_i" to "to_f"

+++ lesson1/triangle_area.rb
@@ -1,10 +1,10 @@
 puts 'OK, now we are going to calculate area of a triangle.'
 
 print 'Enter base: '
-base = gets.strip.to_i
+base = gets.strip.to_f
 
 print 'Enter hight: '
-hight = gets.strip.to_i
+hight = gets.strip.to_f
 
 area = (base / 2) * hight
